Move shared slider method to block folder
